Implemented approach / service request workflows

+++ frontend/src/app/core/header/header.component.ts
@@ -1,20 +1,19 @@
-import { Component, ElementRef, inject, OnInit, ViewChild } from '@angular/core';
+import { Component, inject, OnInit, signal } from '@angular/core';
 import { environment } from '../../../environments/environment';
 import { Router, RouterLink } from '@angular/router';
 import { TranslatePipe, TranslateService } from '@ngx-translate/core';
 import { UtilService } from '../../shared/services/util.service';
 import { UserIdentity } from '../../shared/models/user-identity';
 import { UserRole } from '../../shared/enums/user-role';
-import { ConfirmModalComponent } from '../../shared/components/confirm-modal/confirm-modal.component';
 import { AlertService } from '../../shared/services/alert.service';
 import { AlertType } from '../alert-manager/enums/alert-type';
+import { ApproachService } from '../../features/approaches/services/approach.service';
 
 @Component({
   selector: 'app-header',
   imports: [
     RouterLink,
-    TranslatePipe,
-    ConfirmModalComponent
+    TranslatePipe
   ],
   templateUrl: './header.component.html',
   standalone: true,
@@ -30,8 +29,8 @@ export class HeaderComponent implements OnInit {
   /** Current user identity. */
   protected currentUserIdentity?: UserIdentity;
 
-  /** Logout confirm modal. */
-  @ViewChild("confirmModal") protected confirmModal?: ElementRef;
+  /** The number of pending approaches linked to the current user's ads. */
+  protected pendingApproachCount = signal(0);
 
   protected readonly UserRole = UserRole;
 
@@ -47,25 +46,43 @@ export class HeaderComponent implements OnInit {
 
   private readonly router = inject(Router);
 
+  private readonly approachService = inject(ApproachService);
+
   //endregion
 
   //region methods
 
-  ngOnInit() {
+  async ngOnInit() {
     this.currentUserIdentity = this.utilService.getCurrentUserIdentity() ?? undefined;
+
+    if (this.currentUserIdentity?.roles.includes(UserRole.SERVICE_PROVIDER)) {
+      this.approachService.onPendingApproachCountChanged.subscribe(async () =>
+        this.pendingApproachCount.set(await this.approachService.getUserPendingApproachCountAsync())
+      );
+      this.approachService.onPendingApproachCountChanged.emit(); // Get current user's pending approach count
+    }
   }
 
-  /** Sign out the current user and redirect to the home page. */
-  protected signOut() {
-    this.utilService.removeCurrentUserIdentity();
-    this.currentUserIdentity = undefined;
+  /** Sign out the current user after confirmation and redirect to the home page. */
+  protected async signOutAsync() {
+    this.alertService.showConfirmation(
+      async () => {
 
-    this.router.navigateByUrl('/').then(
-      _ => this.alertService.pushAlert(
-        AlertType.SUCCESS,
-        this.translateService.instant('HEADER.SIGN_OUT_SUCCESS_TEXT'),
-        7
-      )
+        // Remove current user identity
+        this.utilService.removeCurrentUserIdentity();
+        this.currentUserIdentity = undefined;
+
+        // Navigate back to home page
+        await this.router.navigateByUrl('/');
+
+        // Show success message
+        this.alertService.pushAlert(
+          AlertType.SUCCESS,
+          this.translateService.instant('HEADER.SIGN_OUT_SUCCESS_TEXT'),
+          7
+        );
+
+      }, this.translateService.instant('HEADER.CONFIRM_SIGN_OUT_MESSAGE')
     );
   }
